refactor(src/protocol.rs): simplify URL parsing
Non-URL-safe base64 data is no longer supported.

When the origin URL is "https://example.com/abc?", the normal base64 encoding is as follows: "aHR0cHM6Ly9leGFtcGxlLmNvbS8/".

Then the protocol URL is "mpv://play/aHR0cHM6Ly9leGFtcGxlLmNvbS8/" (without parameters and trailing slash), in this case we don't know if the slash is part of the base64 data or the end of the URL.

To avoid this, enforce URL-safe base64.